% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addresses}
\alias{addresses}
\title{Geocoded addresses to the hundred-block}
\format{A data frame with 5 variables: \code{address} (from the original report), \code{lat}, \code{lon} \code{formatted_address} (address returned by API), and
\code{loc_type}. See \href{https://developers.google.com/maps/documentation/geocoding/intro}{Geocoding API}
and \code{?ggmap::geocode()} for more.}
\usage{
addresses
}
\description{
On going list of addresses from the \href{http://opendata.charlottesville.org/datasets/crime-data}{CPD Crime Reports}, that have been successfully geocode via GoogleMaps
API. BlockNumber values of \code{0} or \code{NA} get recoded as `100``, for better accuracy.
Useful for saving that query limit.
}
\keyword{datasets}
