% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{CPD Crime Data}
\format{A data frame with 13 variables, the first 9 are directly from the
the original report, the extra 4 are: \code{lat}, \code{lon}, \code{formatted_address}, and
\code{loc_type}}
\usage{
crime
}
\description{
The last 5 years of requests for police assistance in Charlottesville, Virginia
excludeing homicides,
suicides, and sex offenses. Sourced from the \href{http://opendata.charlottesville.org/datasets/crime-data}{opendata.charalottesville.org}
and geocoded with the GoogleMaps API. Addresses that could not be geocoded, are
included with \code{NA} values in the columns related to geocoding.
}
\keyword{datasets}
