% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_geocode}
\alias{extract_geocode}
\title{Pair down the lengthy JSON object that gets returned from the API}
\usage{
extract_geocode(api_result)
}
\arguments{
\item{api_result}{The return value of \code{ggmap::geocode()}.}
}
\value{
A 1x4 \code{tibble} with snake_case \code{names()}.
}
\description{
Pulls four fields ('geometry.location.lat', 'geometry.location.lon',
'geometry.location_type', 'formatted_address') form the Google Maps API result
into a tibble.
}
