% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{re_geocode}
\alias{re_geocode}
\title{Repeat failed Google Maps API queries}
\usage{
re_geocode(data, retry = 2)
}
\arguments{
\item{data}{A data frame with a column named 'address'}

\item{retry}{An integer for the number of times to re-attempt addresses that
failted to properly geocode.}
}
\value{
The original data frame with new columns 'geocode', a list, of raw API
query results, and 'geocode_good', a boolean, \code{TRUE} if successful.
}
\description{
A persistant function that re-attempts failed address geocode queries, a set # of times or until all
queries are succesfully.
}
